move chat input state into input itself. sep. flow files for sub components to reduce re-render problems
