+++ shared/chat/conversation/input.js.flow
@@ -1,0 +1,14 @@
+// @flow
+import {Component} from 'react'
+
+import type {ConversationIDKey} from '../../constants/chat'
+
+export type Props = {
+  emojiPickerOpen: boolean,
+  isLoading: boolean,
+  onAttach: () => void,
+  onPostMessage: (text: string) => void,
+  selectedConversation: ?ConversationIDKey,
+}
+
+export default class Input extends Component<void, Props, void> { }
